Fixes Type Confusion/Infinity Issue
The type of `amount` was originally assumed to be a number (Gemini docs say 'decimal'). The delivered type is a string, which (when combining trades from a common order) could result in a string representation of an invalid number, potentially larger than MAX_SAFE_INTEGER. This could result in "Infinity" being reported as the amount of BAT purchased by the Brave Team.